102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@ int main(void)
 {
 	int i;
 	int t1 = 1, t2 = 2;
-	int next = t1 + t2;
+	long int next = t1 + t2;
 
 	printf("%d, %d,", t1, t2);
 	for (i = 3; i <= 50; i++)
